refactor: move tax-offset to tools
