--- apps/frontend/lib/disputes/dispute-types.js
@@ -1,4 +1,0 @@
-export default {
-  labels: ["Federal student loan debt", "Credit card debt", "Other"],
-  values: ["federalStudentLoanDebt", "creditCardDebt", "other"],
-};
