+++ apps/frontend/lib/tools/tax-offset/index.js
@@ -1,0 +1,5 @@
+export { default as TaxOffsetAForm, taxOffsetASchema } from "./a";
+export { default as TaxOffsetBForm, taxOffsetBSchema } from "./b";
+export { default as TaxOffsetCForm, taxOffsetCSchema } from "./c";
+export { default as TaxOffsetDForm, taxOffsetDSchema } from "./d";
+export { default as TaxOffsetEForm, taxOffsetESchema } from "./e";
