+++ apps/frontend/pages/disputes/tax/b.js
@@ -1,24 +1,3 @@
-import FieldTemplate from "../../../components/FieldTemplate";
-import Form from "react-jsonschema-form";
-import React from "react";
-import sharedSchema from "../../../lib/disputes/taxes-offset-shared";
+import { TaxOffsetBForm } from "../../../lib/tools";
 
-const schemas = {
-  ...sharedSchema,
-};
-
-const log = type => console.log.bind(console, type);
-
-const DisputeTaxesB = () => (
-  <Form
-    showErrorList={false}
-    FieldTemplate={FieldTemplate}
-    uiSchema={schemas.ui}
-    schema={schemas.json}
-    onChange={log("changed")}
-    onSubmit={log("submitted")}
-    onError={log("errors")}
-  />
-);
-
-export default DisputeTaxesB;
+export default TaxOffsetBForm;
